Fix DbgAssertRefCount on Ctx
Summary:
We don't generate DbgAssertRefCount's for Ctx types, but IncRef and
DecRef can now be applied directly to Ctx, and if refcount-opts
eliminates a pair, it converts them to DbgAssertRefCounts.

I've not seen this happen yet, but with some more diffs I'm working
on, Ctx IncRefs are getting optimized away, and we hit an assertion
because DbgAssertRefCount doesn't support Ctx.

I don't see any reason why it /can't/ happen right now, so putting
this up as a quick fix.

Reviewed By: kodafb

Differential Revision: D3698784

fbshipit-source-id: e92c814fd115ade1b825a4c0b008e9e6610df5ad