Implement variance rules for static methods
Summary:
A type unsoundness was reported through our open source github repo (https://github.com/facebook/hhvm/issues/7271).
We currently allow covariant type parameters in contravariant positions for static methods, and contravariant type parameters in covariant positions. This is a type hole. Consider the case when you emulate polymorphic static method dispatching via instances of classname, as in the repro.

The fix is simply to make the same checks on variance as for non-static methods. Notice that final statics don't need the check, as they can't be overridden.

Reviewed By: alexchow

Differential Revision: D3690524

fbshipit-source-id: 4333318e1a4ed14189f944b5767703a9c024abad