Improvements to VMRegGuard
Summary:
We have a number of VMRegAnchors which are there to protect against
libraries that might be built without frame-pointers, but which can
call back into php. Currently, we have to setup the VMRegAnchor prior
to calling into the library, even though most of the time, it won't be
used.

This is almost exactly the case that VMRegGuard was created for, and
by extending it to capture the current fp, we can use it to cover this
case too.

In addition, by capturing the current fp, when we *do* instantiate a
VMRegAnchor deep within some c++ code, it can skip all the
intermediate frames, and start right at the boundary between c++ and
the jit.

Reviewed By: mxw

Differential Revision: D3693397

fbshipit-source-id: bfab79cd32afa7ca276f16f8e7b6b98b52a60532