Keep a mapping from addresses to their containing VM metadata objects
Summary:
This map will be uesful for profiling memory accesses to persistent
metadata objects like Funcs and Classes.

The current map data structure requires three bits for every eight
bytes in the total range of allocations of objects that we track (modulo
rounding---we round the range to the nearest 2MB, and we round the
number of bits to the nearest 64 bits).  This should result in roughly
a 3/64 memory footprint increase for the tracked object types.

Reviewed By: kodafb

Differential Revision: D3677005

fbshipit-source-id: 3360815426278546cad3db7739e6bf5732a7fb0d