Adding type constraint `where` clause syntax to hhvm's parser
Summary:
Work is underway to implement `where` clauses for generalised type constraints in Hack. This diff adds the required syntax to the HHVM parser, so that the next HHVM release doesn't stumble on this new feature. All the parse results are discarded, because there are no run-time implications to these generalised type constraints.

In other words, it's purely a Hack feature, but HHVM should be able to parse files using it.

Reviewed By: andrewjkennedy

Differential Revision: D3671118

fbshipit-source-id: c8ec58ecbe754c6253284fb4e7a5c201f6a6ece5