Move types from ServerCommand to a separate module, ServerCommandTypes
Summary:
I want to abstract all the main loop logic that interacts with file descriptors into a separate module, which then can be exchangeable.

Right now, this logic is spread in ServerMain and ServerCommand, with the former depending on the latter. Both of those modules will have to depend on my new module, but this module also needs to use the types from ServerCommand, which creates a cyclic dependency. Moving the types to separate module avoids it. I could move it to the new client module, but it doesn't really logically fit there.

Reviewed By: jamesjwu

Differential Revision: D3699946

fbshipit-source-id: c14d67f45063ca7bfd67ffd6661630b1c887a199