Encapsulate main loop handling of file descriptors in ClientProvider
Summary:
ClientProvider is a module with two opaque types, "t" and "client" which hide the fact that we get new clients from a file descriptor and talk to them through a pair of file descriptors.

In a subsequent diff, we can inject another version of this module that will be backed up by different source of clients.

I converted only the code paths that I needed to show a working version of the most basic test - there are still paths left that assume file descriptor and will fail if hit in test. I'll convert them on the as-needed basis.

Reviewed By: dlreeves

Differential Revision: D3700000

fbshipit-source-id: 0209c21307db00ddc921a8182ee2c4eaf92462e3