Remove useless try...with
Summary:
An exception inside this block leaves the server in corrupted state (some classes were redeclared, some were possibly not), and just lead to crash later on during next command that tries to access the same area of shared memory.

The exception will still be caught in ServerMain.handle_connection, and the server state will go to hell, but the latter callsite at least tries to print backtrace, while this function swallows it with premeditation making any debugging a nightmare.

Reviewed By: alexchow

Differential Revision: D3720157

fbshipit-source-id: 48ea2fb0f62840da26b2a586d23ff9f7b9fdeab5