Add test for hfsort
Summary:
Add two tests for hfsort, one for clusterize and one for pettis. Note that the
pettis test is only looking at the number of functions in the resulting hotfuncs
file. I found that the behavior was somewhat implementation specific (in
particular the order of the clusters appears to be mostly unspecified in this
implementation, as well as the order in which they are merged), this presented
problems when trying to use more efficient structures in a later diff.

Reviewed By: kodafb

Differential Revision: D3680354

fbshipit-source-id: 300de801c6c00182dcd51e01fedcbe6356637ed1