Re-organize Datatype values
Summary:
Re-organize the values for Datatype to allow room for KindOfVec,
KindOfPersistentVec, KindOfDict, KindOfPersistentDict, KindOfKeyset, and
KindOfPersistentKeyset. There is not enough space to reserve a separate bit for
both vecs, dicts and keysets, so instead use a different scheme. Reserve a
single bit to represent all Hack arrays and use the remaining free bits as an
enumeration for the Hack array type.

With this scheme, there's no regression in testing existing types, as they all
continue to use their reserved bits. Furthermore, testing for the union of Hack
and PHP arrays is easy, as its just checking if either bit is present (which can
be done with a test instruction. Checking for a specific Hack array type is only
very slightly less efficient. Instead of a single bit test, a mask is applied
and the result is compared against a fixed value (one extra instruction).

The previous Datatype values had a few properties couldn't be preserved with the
new values. All the scalar types were grouped together, meaning IsScalar could
be implemented with a single range check. Now the scalar types are separated, so
IsScalar needs to do several comparisons. However, IsScalar isn't a hot path.

Previously a mask could be used to quickly determine if a Datatype is allowed as
a constant value. However, this is only ever used in slow paths, so use an
explicit switch statement instead.

Finally, the Datatype to destructor table needs to be expanded because the
presence of the Hack array bit causes the entries to be more sparse.

This diff does not actually add KindOfVec, KindOfPersistentVec, KindOfDict,
KindOfPersistentDict, KindOfKeyset, and KindOfPersistentKeyset, but does
illustrate where they'll go.

Reviewed By: paulbiss

Differential Revision: D3489928

fbshipit-source-id: c3943f016a37b9a2a5b30e51f06f2143636220e0