Add KindOf[Persistent]Vec, KindOf[Persistent]Dict, and KindOf[Persistent]Keyset
Summary:
Add the new Datatype kinds KindOfPersistentVec, KindOfVec,
KindOfPersistentDict, KindOfDict, KindOfPersistentKeyset, and
KindOfKeyset. Since we, in general, do not use default cases when we switch on
Datatype, add all the necessary cases everywhere necessary. This is the bare
minimum to get everything compiling, along with a few "obvious" cases.

Start adding rigourous assertions to ensure that vec/dicts/keysets are going
down code paths meant for arrays and vice-versa. Since this diff is not
sufficient for full functionality, this causes the vec and dict test cases to
start failing. Disable the test cases for now until the full support is
completed, where they will be re-enabled.

allow-large-files

Reviewed By: paulbiss

Differential Revision: D3489952

fbshipit-source-id: 7e2a770fd5a962ca4e717daab0856618ce1ea1f8