Remove DefineReffy
Summary:
Remove the last usages of the DefineReffy MOpFlags. It was ignored in almost all
cases, except for vecs and dicts, and collections in one place. In fact, vecs
and dicts were using it incorrectly. The intention is to forbid ref binding in
the final member operation, but the DefineReffy flag is set for all intermediate
operations if the final operation is a bind. As a result, one is unable to bind
a ref inside an array which is inside a vec or dict.

Instead of DefineReffy, provide an explicit bool template parameter in the few
places where the distinction matters. It is only provided for final operations.

Reviewed By: paulbiss

Differential Revision: D3489974

fbshipit-source-id: 197f54f3245164c503c96df0f7870f2c75af662e