Add support for vecs/dicts/keysets to APC
Summary:
Add support for storing vecs, dicts, and keysets to APC. Support is
very similar to that of arrays, except there is no equivalent of
APCLocalArray. Since vecs/dicts/keysets do not currently support
specializations, it is not possible to have a specialization which references
the APC stored data. Instead, the data is turned into a normal vec/dict/keyset
on demand.

Reviewed By: kodafb

Differential Revision: D3489988

fbshipit-source-id: 658ee938d4e90524917fac491549942e6f495e07