Fix potential mem leaks when converting to vecs and dicts
Summary:
When converting an array to a vec or dict, an exception should be thrown if the
source array contains a reference. However, in certain cases, the new vec/dict
may not be freed properly when the exception is thrown.

This can be fixed by scanning the input array for references first before
allocating anything. The result of this scan can have three possible results. If
there's a reference which has a ref-count greater than 1, than just throw. If
there's no references, proceed with the conversion (with the added benefit that
we can potentially do the conversion without copying). If there's a reference
with has a ref-count equal to 1, we can proceed, but we have to copy (this will
collapse the ref as a side-effect).

While we're at it, add more assertions for sanity checking, and allow the caller
of the conversion to specify whether a copy is needed or not.

Reviewed By: paulbiss

Differential Revision: D3490035

fbshipit-source-id: a0c688bc7d86cbeb54cd9ceb178708dc4c149b85