Re-work array comparisons and add support for vec/dict/keyset
Summary:
Re-work array comparison to handle vecs, dicts, and keysets
easier. Rather than have the comparison functions be member functions of
ArrayData, instead provide static functions which perform the comparison. The
functions within ArrayData are for normal PHP arrays, while PackedArray and
MixedArray provide versions for vecs and dicts/keysets respectively. The member
functions within ArrayData simply determine whether the ArrayData* represents a
PHP array, vec, dict, or keyset (throwing if there's a parameter mismatch), and
call out to the right static function.

As an advantage of this, the JIT can call the appropriate static function
directly, avoiding the overhead of checking whether the ArrayData* is a PHP
array, vec, or dict.

This diff only adds the support for vec/dict/keyset comparisons in the runtime,
which the interpreter will use. The JIT support will come in a later diff.

Reviewed By: paulbiss

Differential Revision: D3490058

fbshipit-source-id: 5ba389e84be8c10bddcb58de539d6e4c03eb38c0