Add misc JIT support for vec/dict/keyset
Summary:
Add some miscellaneous JIT support for vecs, dicts, and keysets which
didn't belong to other diffs. Add opcode CountDict and CountKeyset which returns
the size of a dict or keyset, and use them in a few places. Also add support for
iterators and check for TArrLike instead of TArr in a few places.

Reviewed By: paulbiss

Differential Revision: D3490236

fbshipit-source-id: ed0a8314916fa7d3bdfe8897613ea72a093479aa