Exactly scan "normal" RDS during GC using type-scanners
Summary:
Rather than conservatively scanning the entirety of RDS during garbage
collection, exactly scan the "normal" section using type-scanners. The
"persistent" section is shared amongst all threads, so it should never contain
pointers to request allocated memory, so it doesn't need to be scanned at
all. The "local" section can contain pointers to request allocated memory, but
it is only initialized under certain circumstances, so continue to conservative
scan it for now.

This required a new type-scanner marker MarkScannableCountable<>, which
indicates that a type is countable, but there should be a scanner generated if
explicitly requested. This is needed because RefData is unique amongst countable
types in that it can exist outside of the request heap (such as RDS or the
stack).

Reviewed By: edwinsmith

Differential Revision: D3558202

fbshipit-source-id: bbe748846bd5108c4a0e8c2c53c9e50eddac5793