Add additional dependencies when overriding an element
Summary:
In {D3564703} I changed it so the types of class elements are stored in the separate part of shared memory. This requried changing what dependencies we track for incremental type checking in {D3589777}. While this caught most cases where we add dependencies, there were some places where we fetch a member of a class that doesn't go through `Typing_env`, in particular in `Typing_extends`. `Typing_extends` makes sure a child class is compatible with its ancestors. This requires looking up the type of a class element, but we weren't recording a dependency when we do so.  This diff adds the necessary dependencies.

For a repro of the issue consider the following classes, each declared in another file.

Reviewed By: dabek

Differential Revision: D3732957

fbshipit-source-id: 2a7159a8422dfbbe9aead8f59f7b39f13b8b10fa