Move static/instance method naming conflict check to typing phase
Summary:
During an incremental type check, the contents of the `Decl_heap` can be left in an inconsistent state. For instance imagine the following code where each class is in its own file.

```
<?hh // strict
abstract class Base {
  public static function meth(): void {}
}

<?hh // strict
abstract class Parent extends Base {
}

<?hh // strict
class AChild extends Parent {
  public static function test(): void {
    $achild = new self();
    $achild->__meth();
  }

  private function __meth(): void {}
}
```

Now lets say we get the following changes in a batch.

```
<?hh // strict
abstract class Base {
}

<?hh // strict
class AChild extends Parent {
  public static function test(): void {
    $achild = new self();
    $achild->meth();
  }

  private function meth(): void {}
}
```

Two changes happen at the same time, we are removing the static method `Base::meth()` and we are renaming the instance method `AChild::__meth()` to `AChild::meth()`.

When we see these changes come in we start by oldifying both `Base` and `AChild` in the `Decl_heap` and then declare both classes. When we declare `AChild` first, we do so based on inconsistent state. `Parent` still has the members folded from `Base` previous definition, so it contains a reference to the static method `Base::meth()`. `AChild` now thinks it inherits `Base::meth()` which conflicts with its own `AChild::meth()`. When it tries to find the position of `Base::meth()` to report an error, none is available because `Base::meth()` has already been moved to the old heap, so the error will have `Pos.none`. This then triggers an assertion in `Decl_redecl_service` that crashes the server.

Note that this is only inconsistent for the first phase of redeclaration. For that first phase we ignore all errors that are reported and end up redeclaring `AChild` anyways (though we may end up retype checking things unnecessarily).

To fix this particular case I moved the check to the typing phase, when everything should be settled. In general though we may want to reconsider doing error reporting during declaration because of the inconsistent state redeclaration leaves the world temporarily.

Reviewed By: dabek

Differential Revision: D3734508

fbshipit-source-id: 5dffd0d4ff9d313d9ac0fbdbd90db33777303ba8