Fix repo test failure on unsetting vec elements
Summary:
Unsetting elements in a vec now only might turn it into a dict (if you
remove the last element, or remove a non-existent element, it is
unchanged). However, hhbbc was still assuming that it always turned into a
dict. This caused repo-auth-type assertion failures in a test case.

Also add a slight pessimization to minstr-effects for unsetting on a vec, as it
has exposed an issue with how we dec-ref known counted array types. This will be
fixed in a later diff.

Reviewed By: mxw

Differential Revision: D3732905

fbshipit-source-id: 09ec62f7d2e68276d2dc620b46aa35994ed5e0f2