Prevent calling closures statically
Summary:
Invoking a closure statically crashes hhvm (we assume its always invoked on
a closure object).

One simple fix is just to add a check in the closure prolog, but that
pessimizes every closure call.

It turns out that zend php fatals for this case, and further digging
reveals that it fatals for builtin methods that expect a $this but
don't get one, and also for "special" class methods - ##__construct,
__destruct, __invoke## etc. In addition, even for normal methods it
raises a notice.

Their fatal/notice is raised before the arguments are evaluated - so
putting the fix in the closure prolog seems wrong.

This adds support for doing the same in hhvm. Initially, it enforces
the fatals, but only raises notices under an option which defaults to
false when HipHopSyntax is enabled.

I'm hoping this is pretty much free, and we can experiment with enabling
the notice always later.

Lots of unit tests failed due to notices - so for now I've updated the
config to turn them off. A couple failed due to fatals, and in those
cases I updated the tests to work as I think was intended.

Reviewed By: paulbiss

Differential Revision: D3699631

fbshipit-source-id: ae5c17185f3730dcee518d0f91dfd211af553926