Report number of nonempty slots in SharedMem.hash_tbl_stats
Summary:
We leak hash table slots by not setting the slot's hash to `0` after removing elements. For now I want to log this difference to see if its related at all to some crashes we are seeing before trying to fix it.

Also print our the hash table stats when exiting due to running out of shared memory or running out of hash table slots

Reviewed By: dabek

Differential Revision: D3739041

fbshipit-source-id: acb1bc4652f39bcdfabaee2300dbc315f5bf43f2