Add full-fidelity support for parsing late static bindings
Summary:
The full-fidelity parser currently parses all occurrences of
the Static token as leading off a static declaration, though
"static" can also be used to designate a late-bound scope for
a property access or method invocation. (E.g., "static::foo" or
"static::bar()", or Class_const and Class_get nodes in the
current AST.)

These changes modify the full-fidelity parser to disambiguate
between uses of "static" in decl and expression contexts by
checking for the presence of a ColonColon token following the
Static token. Uses of "static::..." will now parse as an
expression statement.

Reviewed By: jamesjwu, eshrews

Differential Revision: D3693172

fbshipit-source-id: 0d9112d8b4b10bedd3a69a7f09556387a7da3a9c