Restrict contravariant uses of 'this' in final classes
Summary:
Under normal circumstances, the `this` type is treated as bivariant, with flow-sensitive checking determining when an access violation off
of a `this`-typed entity occurs.
Some of these checks are relaxed in the presence of a final class, since `this` can only indicate the enclosing runtime type, but doing so
leaves room for a soundness hole if `this` is referenced contravariantly. For example:

```
class Base {}
class Derived extends Base {
  public function __construct(public int $v) {}
}

final class ImplCov<+T> {
  public function __construct(private T $v) {}
  public function put(this $v): void {
    $this->v = $v->v;
  }
  public function pull(): T {
    return $this->v;
  }
}

class Violate {
  public static function foo(ImplCov<Derived> $v): void {
    self::bar($v);
    echo $v->pull()->v; // Wait... Base doesn't have $v!
  }
  public static function bar(ImplCov<Base> $v): void {
    $v->put(new ImplCov(new Base()));
  }
}

function BreakIt(): void {
  Violate::foo(new ImplCov(new Derived(53)));
}
```

In the code above, the covariance of `T` will allow for assignments from `ImplCov<Derived>` to `ImplCov<Base>` in the argument position.
Unfortunately, doing so leads to a runtime crash if `BreakIt` is invoked, per the assignment of an `ImplCov<Base>` object into `$v`, which
was delared as having type `this`.

To address this problem, it was decided to restrict the contravariant use of `this` in variant final classes, because in those (very narrow)
cases a guarantee can't be made as to what type `this` will be at runtime. These changes update the variance checker to diagnose such instances
of `this`-typed arguments and constraints.

Reviewed By: dlreeves

Differential Revision: D3737899

fbshipit-source-id: f9bb60baeed8e0cee5a10f3e38de3221a998dc0a