Replace kFuncCountHint with two different runtime options
Summary:
I've added Eval.FuncCountHint and Eval.PGOFuncCountHint, hints for the
total number of Funcs and PGO-involved Funcs, respectively. The AtomicVectors
and AtomicHashMaps that were sized based on kFuncCount are now sized based on
the new options. Some of the AtomicVectors are constructed at process startup
time, so I added a way to keep track of them and avoid fully constructing them
until we've loaded all runtime options.

Reviewed By: ricklavoie

Differential Revision: D3732232

fbshipit-source-id: 1353c5067cc09606e71584e75b6c2dd4db382b89