Add a mechanism to log "perf warnings"
Summary:
There are a number of different things that can happen in hhvm that
aren't fatal but can cause reduced performance. This diff adds a way for us to
log these events somewhere, rather than choosing between crashing and ignoring
them until someone manually checks for each one.

This diff includes two events to get things started: AtomicHashMap submap
allocation and vasm-xls punts.  Only 1% of events are logged by default, with
the ability to override this both globally (with Eval.PerfWarningSampleRate)
and per-event (with an argument to logPerfWarning()).

Reviewed By: kodafb

Differential Revision: D3419912

fbshipit-source-id: a47f63d4379030d9a8b6eca3749a8b327652207e