Fix unused function warning in gen-member-reflection
Summary:
When HHVM is built without gen-member-reflection enabled (IE, debug
builds), there's no actual call to generate_entry(), which causes an "unused
function" warning to be emitted. Fix this by not removing the entire generation
loop using the preprocessor, but instead checking a constexpr bool (which is set
according to the preprocessor).

Reviewed By: mxw

Differential Revision: D3751714

fbshipit-source-id: 4fc056cae4d0328b817bfb14cc0e3cdd1cad8248