Moved integration tests to new folder and added new tests
Summary:
A first attempt at moving some of our integration tests over to Kasper's new framework.
So far, these tests will run with a fresh init of the server, and have a few differences from the python tests (in config, etc).
We'll probably want to figure out a way to run these tests with a saved state, etc, later.

Reviewed By: dabek

Differential Revision: D3754181

fbshipit-source-id: 9fad2d4b1261ebd3236c11a0ea59f40d7de5e400