Create "server_command_types" TARGET that is independent of "server"
Summary:
In order to mock client sending commands in integration tests, I need to be able to talk about command types without talking about ServerEnv (cyclic dependencies occur otherwise).

This diff creates a new build target with just those types that doesn't depend on ServerEnv. The main change of this diff is moving ServerRpc.t type, which is the cause of most cycles, to ServerCommandTypes.

Reviewed By: dlreeves

Differential Revision: D3724557

fbshipit-source-id: 666313c244a9d098a58e1ca1f3b0addc2b402ccc