Simplify rechecking logic of IDE files
Summary:
Right now, every file has "last edit time", and we recheck only those files that have not been modified in last 0.5 seconds (files that are not "being edited").

Tracking it on a per-file basis is not really necessary here - we can greatly simplify it by only tracking the time of last edit command (to any file), and based on that do a recheck of ALL or NONE of the IDE files.

Reviewed By: dlreeves

Differential Revision: D3757798

fbshipit-source-id: 0f96b0e87bcd4700e06eb3a02687414c7836f19a