Use Relative_path instead of strings in IDE code
Summary:
There is a lot of trashing and converting between strings and paths because IDE code uses ones, and pre-existing typechecker code the latter. Converting everything to use Relative_paths as long as possible (i.e. most often until we need to send it to client/actually read it from the disk).

The main change is in serverEnv.ml, the rest follows from it.

Reviewed By: dlreeves

Differential Revision: D3757802

fbshipit-source-id: eff7bec88abd4a79436fa8953a5bdd093da32637