Fix vasm simplication of "cmp 0" to "test" for PPC64
This improvement implemented on fa442471676c9550691c28fe31c9fee2edda0a05
is arch-specific (does test really runs faster on your arch than cmp?).
Moreover it also assumes that your arch can emit the vasm "test",
whereas PPC64 can't. It lowers it to something else (mostly due to the
need for temporary registers) so this simplification is breaking most of
the tests.

This is a quick and dirty fix. Maybe a refactor for having arch-specific
simplifications are actually a better approach.