Fix rank violation in hphp_compiler_parse
Summary:
When hphp_compiler_parse fails to write a unit into the repo it will try to
reconstruct the unit using the contents of the repo (in order to handle races
with other threads). When this happens the unit that was previously intended for
storage in the repo is discarded.

Destructing a unit while holding the tbb mutex for the non-repo unit cache is
problematic in recycle-tc mode. When the unit is destructed recycle-tc will
try to free any functions it may contain. In order to do this the code and
metadata locks must be acquired, leading to a rare rank violation assertion in
debug builds.

Reviewed By: mxw

Differential Revision: D3762713

fbshipit-source-id: d9adcda80c9d75d6d521a2d06e6b95332e5fa47b