Remove type from DecRefProfile key
Summary:
DecRefProfile is used to try to estimate whether a particular DecRef is likely
to result in a release. Part of the key used to identify particular profiler
instances is the type of the SSATmp being dec-refed. However, the type of a
SSATmp can change from profiling translations to optimized translations
(particularily if type predictions are inserted). As a result, we can lose the
profiling information about some DecRefs in the process. Remove the type
component of the key to avoid this loss of data.

Reviewed By: ottoni

Differential Revision: D3758581

fbshipit-source-id: 5596a387424c4373f1e226d9ddd8dbef93210ada