Allow "noreturn" as a generic type argument
Summary: The spec indicates that "noreturn" is only legal as a return type hint, but plainly this is wrong because Awaitable<noreturn> is legal. I will figure out what the right language should be in the specification later; for now, we allow noreturn to appear in a generic type argument list. (Likely we will also need a later error detection pass to determine when noreturn is being used improperly; List<noreturn> should probably be illegal, for instance.)

Reviewed By: dlreeves

Differential Revision: D3718622

fbshipit-source-id: 366733e3d94af7a62705fc030ce8b44242a35df4