Fix cmp to test vasm simplication for PPC64
Summary:
The improvement implemented on fa442471676c9550691c28fe31c9fee2edda0a05 is
arch-specific as a test may not be better than a cmp on every arch. So it
should actually be opt-in for other archs to make use of it.

Moreover PPC64 couldn't handle the test vasm after it's lower phase, therefore
a second call to lower was added to guarantee that the vasms emitted by
simplify and optimizeCopies are actually handled by PPC64 emitter.

No significant overhead was noticed.
Closes https://github.com/facebook/hhvm/pull/7289

Reviewed By: mxw, Orvid

Differential Revision: D3765305

fbshipit-source-id: be49d4bf36da4125c324b66b03954a6e19af363c