Allow unification of covariant 'this' in final classes
Summary:
My original changes to unification in D3737899 were too restrictive, but I didn't
realize it because I wasn't seeing errors from my www builds.

This change slightly relaxes one of the checks I put in place, allowing for the
unification of covariant `this` types in final classes. This is technically safe,
since it only affects methods on variant classes that return a covariant `this`-
typed, via a return expression that is also of a covariant `this` type. (See
flib/entity/entql/core/EntQLIDQuery.php for one such case.)

Reviewed By: dabek

Differential Revision: D3775059

fbshipit-source-id: 6bee8e3678da74d238e291453fada95329b71f3e