Create nodes for member selection (?->, ->) and scope (::) operators
Summary:
Member selection and scope resolution operators are sufficiently unusual operators to warrant their own special tree nodes and data structures; they are not quite like the run-of-the-mill binary operators.

This commit just creates the data structures and pretty printer; the actual parsing rules will be added in later diffs.

Reviewed By: jopamer

Differential Revision: D3753865

fbshipit-source-id: 31943513027b9d9f4cf80ea23ff682a2c9f5523e