Move precedence detection code to better place
Summary:
We must check whether the operator which follows a given expression is of lower precedence than operators in the already-parsed expression. But we were only consistently doing that in the code which handles binary operator expressions; its better to have that check in the code which handles all "is there something after this expression?" code.

Surprisingly, this does not make a difference in any test cases. However, it will be necessary for future diffs, when I add helper methods for parsing -> and :: specially.

Reviewed By: jopamer

Differential Revision: D3754854

fbshipit-source-id: ad81aa880f62c4e5e0eb3b0a2b7c5c600d054bd1