Add 2 more basic integration tests
Summary:
From what I can tell, these are the remaining integration tests from common_tests.py and heirarchy_tests.py that can be easily converted to the new format.
Some needed features to convert the rest are:
- loading from save state
- calling server command flags
- Accessing data in the middle of a typechecker run (files to be redeclared, etc).

I'll investigate those next; I personally feel like some of the tests we have are actually not helpful in their current design, and could probably be rewritten to be more
black box-like, which also would make their conversion easier.

Another thing to note is that I'm already starting to feel the pain of having to type out every error message into ocaml string format, so it might be helpful as a side
to get some sort of test_errors.exp working, as well. I'll think more about how to do this, since unlike in the full fidelity parser, a test does not have a one-to-one relationship with errors or files.

Reviewed By: dlreeves

Differential Revision: D3761467

fbshipit-source-id: e612cb929a5d7b7762db429a0665da558b2d62be