Fix a bug setting up the context for dynamic functions
Summary:
The FuncCache code was written when hhvm (and php) didn't support

  $f = "Cls::meth";
  $f();

so assumed that the called function couldn't have a non-null ctx.

It looks like it was half updated so that it could /find/ such
functions, but not so that it filled in the this/cls field of the
ActRec. This sort-of-worked in simple cases where the called function
didn't inspect the ActRec, but it was easy to setup examples that
crashed hard (eg any of the calls in the new test case).

For now, I disabled caching of methods. I considered augmenting the
cache to hold an additional ctx field - but there are various issues
with doing that (and I've constructed the test case to excercise all
the edge cases I could think of). We can try to re-enable the caching
later, with suitable checks on its validity.

Reviewed By: swtaarrs

Differential Revision: D3782249

fbshipit-source-id: 0774a201be0b64d1e9cea773b97034a2b5a8b799