GetK operations with invalid offsets should be unreachable
Summary:
The simplifier attempts to simplify away MixedArrayGetK, DictGetK, and
KeysetGetK operations where the array is a constant. However, just because one
of those instructions is present, doesn't mean that it is safe to access an
element at the specified offset on the constant array. This is because, of
course, each of those instructions should be guarded by an appropriate check
instruction, and the GetK operation might actually be unreachable. Rather than
blindly trying to access an element at the specified offset, check first if the
offset is valid. If it is not, that's an indication that this instruction is
actually unreachable, and return Bottom instead.

While we're at it, simplify the CheckOffset operations as well.

Reviewed By: paulbiss

Differential Revision: D3784670

fbshipit-source-id: 1fa6a2ebce727f02d205f4e132801f5c83193a1d