Don't manipulate the stack until we're sure we're going to inline
Summary:
As part of preparation for inlining, we first pop off of the stack all of the
arguments to the callee. This values will later be set as locals inside the
callee's inlined code. In the process of inlining, we may detect that we're
calling a method T1:foo, but the object actually has type T2. This can happen
due to inconsistent type assertions and indicates that the call is actually
unreachable. If this occurs, we bail out of inlining and leave the call as
is. However, at this point we've already popped off the arguments and make no
effort to restore them to the stack.

So, at this point, the caller's frame state has been updated to reflect the
argument popping, but it still believes the stack will be popped upon returning
from the call. IE, if there are 3 arguments, the stack should have at least
depth 6 (3 args, and 3 for the pre-live ActRec). Because of the argument popping
in the inliner, we've popped off 3. The frame state manager will assume that
after the call, 6 slots will be popped, meaning the stack has now underflowed by
3. This can cause such effects as "popping" locals, iterators, and ActRecs.

This may not seem necessarily fatal since the code is unreachable. However,
unreachable code can join reachable code at a merge point, and the frame state
from the unreachable code can be carried forward. As a result, we now have
reachable code which has an incorrect knowledge of the stack depth.

Fix all of this by only popping the callee's arguments once we're sure we can
inline safely. IE, do the inconsistent call check first.

Reviewed By: markw65

Differential Revision: D3784705

fbshipit-source-id: 17e9ca4a57c9df58ac23c31693c89344ab4a65b5