Fix a stack overflow issue
Summary:
When we detect a stack overflow, we don't know how much space is left,
so we can't finish setting up the function that caused the overflow, but
its not in the right state for unwinding.

We tried to work around that by setting up the vm state as if we were just about
to call the function, and rely on the caller's fpi regions to sort it all out.

But if our callee was inlined, and its frame elided, we would set
things up as if we were about to call our caller's caller instead; and
that knows nothing about the actrec of the function that ultimately
overflowed the stack.

To fix this, we throw an exception. By the time we catch the
exception, all catch traces have run, the inlined function's stack
frame has been materialized and linked in, and its now safe to set
things up as before.

Reviewed By: mxw

Differential Revision: D3785698

fbshipit-source-id: 61935ac33d0663e1ef47d4113d533a93721a958c