Parse scope resolution (::) to custom parse tree node
Summary:
The :: operator requires a name or "class" on the right hand side; we detect this and give an error if it is wrong.  We then create a parse tree node specifically for scope resolution, rather than treating :: as a binary operator.

Note that this change causes "parent" to become a keyword that is only meaningful in the context of a "new" or "::" expression, which means that the legal code

    $y = $x->parent;

is now an error; the "parent_var" test case now fails. This problem will be addressed in a later diff which changes how "->" is parsed.

Reviewed By: jopamer

Differential Revision: D3758037

fbshipit-source-id: ba75540e182ab431d3dacfa5bf97368770ca4627