Optimize LdARFuncPtr and CheckRefs
Summary:
Mostly hphpc/hhbbc eliminate function parameter reffiness checks. But
generally they don't for closures, even when the closure is defined
locally.

This can lead to odd situations where FPushFuncD pushes a known Func*,
LdARFuncPtr then loads the Func* from the ActRec, passes it to
CheckRefs to do reffiness checks against the params, and then we
inline the body...

This diff allows us to resolve LdARFuncPtr to a known Func* when the
corresponding SpillFrame knew the Func*, and then optimizes CheckRefs
when the Func* is known.

Reviewed By: paulbiss

Differential Revision: D3789818

fbshipit-source-id: c8cbd05a00305eef50ae4193229b965fedcc04cc