Use NativeNode instead of ResumableNode for AsyncFuncWaitHandle
Summary:
This eliminates struct ResumableNode, but preserves the special
HeaderKind::AsyncFuncFrame, with memory layout as follows:

  [NativeNode][frame slots][Resumable][c_AsyncFuncWaitHandle]

NativeNode already provides an obj_offset field, which we can
use to locate either the AFWH object or resumable using simpler
calculations than we did before.

Note that AFWH is still not HNI. Generator and AsyncGenerator
have similar layouts, but *are* HNI, and use a NativeNode with
HeaderKind::NativeData.

Differential Revision: D3777640

fbshipit-source-id: ee7ce8f35da657d3f4e8cbee25b3f2582dac8969