Parse member selection
Summary:
Member selection (-> or ?->) gets its own special parse tree node and parser rules.

The right hand side can only be a variable or a name. Moreover, if a name, it can be a keyword.

In this diff I also update the copy of the grammar used by Shiyu's "random legal program generator" that we use for testing.

Reviewed By: jopamer

Differential Revision: D3759404

fbshipit-source-id: b5bd9ce28f67cf996a33f12b201aece96f639c0c